requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asttokens==2.2.1
backcall==0.2.0
bcrypt==4.2.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
comm==0.1.3
debugpy==1.6.6
decorator==5.1.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
executing==1.2.0
fastapi==0.115.6
Flask==3.0.3
h11==0.14.0
idna==3.10
ipykernel==6.22.0
ipython==8.12.0
itsdangerous==2.2.0
jedi==0.18.2
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.1.0
jupyter_core==5.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.26.4
packaging==23.0
pandas==2.2.2
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.6.1
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.14.0
pymongo==4.10.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.1
pywin32==306
pyzmq==25.0.2
rsa==4.9
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
stack-data==0.6.2
starlette==0.41.3
threadpoolctl==3.5.0
tornado==6.2
traitlets==5.9.0
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.34.0
vboxapi==1.0
wcwidth==0.2.6
Werkzeug==3.0.3
